# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Epistemic Transparency \u0026 Event Lineage System API\nAn append-only event store with hash-chaining for AI decision tracking",
        "title": "Lineage API",
        "contact": {
            "name": "Lineage Support",
            "url": "https://github.com/vkrishna03/Lineage"
        },
        "license": {
            "name": "Elastic License 2.0",
            "url": "https://www.elastic.co/licensing/elastic-license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/actors": {
            "get": {
                "description": "List all actors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "List actors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/actor.ListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new actor (human, llm, agent, service, or tool)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Create an actor",
                "parameters": [
                    {
                        "description": "Actor creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/actor.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/actor.ActorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/actors/{id}": {
            "get": {
                "description": "Get an actor by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Get an actor",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Actor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/actor.ActorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/events": {
            "get": {
                "description": "List all events in a scope, ordered by scope_sequence",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List events by scope",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Scope ID",
                        "name": "scope_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/event.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Submit an event to the processing queue. Events are processed asynchronously and will be hash-chained.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create an event",
                "parameters": [
                    {
                        "description": "Event creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/event.Input"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/event.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "description": "Get an event by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get an event",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/event.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/events/{id}/lineage": {
            "get": {
                "description": "Get parent and child events for a given event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get event lineage",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/event.LineageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "actor.ActorResponse": {
            "description": "Actor data returned by the API",
            "type": "object",
            "properties": {
                "external_id": {
                    "type": "string",
                    "example": "claude-opus-4"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "metadata": {
                    "type": "object"
                },
                "name": {
                    "type": "string",
                    "example": "Claude"
                },
                "registered_at": {
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "type": {
                    "type": "string",
                    "example": "llm"
                }
            }
        },
        "actor.CreateRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "external_id": {
                    "type": "string",
                    "example": "claude-opus-4"
                },
                "metadata": {
                    "type": "object"
                },
                "name": {
                    "type": "string",
                    "example": "Claude"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "human",
                        "llm",
                        "agent",
                        "service",
                        "tool"
                    ],
                    "example": "llm"
                }
            }
        },
        "actor.ListResponse": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/actor.ActorResponse"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "event.CreateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Event queued for processing"
                },
                "status": {
                    "type": "string",
                    "example": "accepted"
                }
            }
        },
        "event.EventResponse": {
            "description": "Event data returned by the API",
            "type": "object",
            "properties": {
                "actor_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                },
                "correction_type": {
                    "type": "string",
                    "example": "supersede"
                },
                "corrects_event_id": {
                    "type": "string"
                },
                "decided_at": {
                    "type": "string",
                    "example": "2024-01-15T10:35:00Z"
                },
                "event_hash": {
                    "type": "string",
                    "example": "def456..."
                },
                "event_type_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440003"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "ingested_at": {
                    "type": "string",
                    "example": "2024-01-15T10:40:00Z"
                },
                "intent": {
                    "type": "string",
                    "example": "decision"
                },
                "observed_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "payload": {},
                "prev_event_hash": {
                    "type": "string",
                    "example": "abc123..."
                },
                "reason": {
                    "type": "string",
                    "example": "User approved the recommendation"
                },
                "scope_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "scope_sequence": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "event.Input": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "correction_type": {
                    "type": "string",
                    "enum": [
                        "supersede",
                        "amend",
                        "retract"
                    ],
                    "example": "supersede"
                },
                "corrects_event_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440003"
                },
                "decided_at": {
                    "type": "string",
                    "example": "2024-01-15T10:35:00Z"
                },
                "event_type_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                },
                "intent": {
                    "type": "string",
                    "enum": [
                        "exploration",
                        "suggestion",
                        "assertion",
                        "decision",
                        "execution"
                    ],
                    "example": "decision"
                },
                "observed_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "parent_event_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "payload": {
                    "type": "object"
                },
                "reason": {
                    "type": "string",
                    "example": "User approved the recommendation"
                },
                "scope_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "event.LineageResponse": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/event.EventResponse"
                    }
                },
                "event_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "parents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/event.EventResponse"
                    }
                }
            }
        },
        "event.ListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/event.EventResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}